requirements.txt
Flask
Flask-SQLAlchemy
mysql-connector-python
pandas
numpy
scikit-learn
matplotlib
seaborn
